requirements.txt
Flask==2.3.3
pandas==2.1.0
plotly==5.17.0
scikit-learn==1.3.0
numpy==1.24.3
python-dotenv==1.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
gunicorn==21.2.0
Werkzeug==2.3.7